Added FPGA-kernel-name-restriction check
Based on Guideline #4 from the "Guidelines for Naming the Kernel" in the Intel FPGA SDK for OPenCL Programming Guide
Source: https://www.altera.com/documentation/mwh1391807965224.html#ewa1411503895045

Matches on FunctionDeclarations, and checks whether or not the FileEntry for the source file in which the function is
declared ends with "/kernel.cl".

In order to not point towards the function itself, a SourceLocation is created for the kernel FileEntry line 1 column 1,
and passed to the diag() function.