[clang-tidy] new altera struct pack align check
Note: this is an update to "new FPGA struct pack align check"; it's been moved to the new altera
module, because the recommendation is taken from the best practices guide for the Altera FPGA
compiler.

This lint check is a part of the FLOCL (FPGA Linters for OpenCL) project out of the Synergy Lab at
Virginia Tech.

FLOCL is a set of lint checks aimed at FPGA developers who write code in OpenCL.

The altera struct pack align lint check finds structs that are inefficiently packed or aligned and
recommends packing/aligning of the structs using the packed and aligned attributes as needed.