[clang-tidy] new altera single work item barrier check
This lint check is a part of the FLOCL (FPGA Linters for OpenCL) project
out of the Synergy Lab at Virginia Tech.

FLOCL is a set of lint checks aimed at FPGA developers who write code in
OpenCL.

Finds OpenCL kernel functions that call a barrier function but do not
call an ID function. These kernel functions will be treated as single
work-item kernels, which could be inefficient or lead to errors.

Based on the "Altera SDK for OpenCL: Best Practices Guide."